backend/requirements.txt
argcomplete==3.6.3
asgiref==3.11.1
boto3==1.42.83
botocore==1.42.83
certifi==2026.2.25
cffi==2.0.0
cfn-flip==1.3.0
charset-normalizer==3.4.7
click==8.3.2
cryptography==46.0.6
defusedxml==0.8.0rc2
dj-database-url==3.1.2
Django==5.1.4
django-cors-headers==4.6.0
django-extensions==4.1
django-filter==25.1
django-formtools==2.5.1
django-guardian==3.0.3
django-oauth-toolkit==3.0.1
django-otp==1.6.1
django-phonenumber-field==8.1.0
django-rest-framework-social-oauth2==1.2.0
django-simple-history==3.8.0
django-storages==1.14.6
django-two-factor-auth==1.18.1
django-unfold==0.55.2
djangorestframework==3.17.1
djangorestframework_simplejwt==5.4.0
drf-social-oauth2==3.1.0
durationpy==0.10
gunicorn==25.3.0
hjson==3.1.0
idna==3.11
jmespath==1.1.0
jwcrypto==1.5.6
MarkupSafe==3.0.3
pymysql==1.1.0
oauthlib==3.3.1
packaging==26.0
phonenumbers==9.0.9
pillow==11.3.0
placebo==0.9.0
psycopg2-binary==2.9.11
pycparser==3.0
PyJWT==2.12.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-slugify==8.0.4
python3-openid==3.2.0
PyYAML==6.0.3
qrcode==8.2
requests==2.33.1
requests-oauthlib==2.0.0
s3transfer==0.16.0
setuptools==82.0.1
six==1.17.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.3
text-unidecode==1.3
toml==0.10.2
tqdm==4.67.3
troposphere==4.10.1
typing_extensions==4.15.0
urllib3==2.6.3
Werkzeug==3.1.8
wheel==0.46.3
whitenoise==6.12.0
zappa==0.62.1
vercel-django